requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-iframetoolbox==0.2.1
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2020.4
sqlparse==0.4.1